:ambulance: :: Long -> Intger 다 변경

+++ src/main/java/com/example/msgadminapi/controller/ClubController.java
@@ -36,13 +36,13 @@ public class ClubController {
     }
 
     @DeleteMapping("{clubIdx}")
-    public CommonResultResponse clubDelete(@PathVariable Long clubIdx) throws Exception {
+    public CommonResultResponse clubDelete(@PathVariable Integer clubIdx) throws Exception {
         clubService.clubDelete(clubIdx);
         return responseService.getSuccessResult();
     }
 
     @PatchMapping("{clubIdx}/end")
-    public CommonResultResponse clubFinishOpen(@PathVariable Long clubIdx) throws Exception {
+    public CommonResultResponse clubFinishOpen(@PathVariable Integer clubIdx) throws Exception {
         clubService.clubClose(clubIdx);
         return responseService.getSuccessResult();
     }
