+++ src/main/java/com/example/msgadminapi/domain/entity/afterschool/Grade.java
@@ -13,14 +13,14 @@ import javax.persistence.*;
 public class Grade {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private Long id;
+    private Integer id;
 
     @JsonIgnore
     @ManyToOne(fetch = FetchType.EAGER)
     @JoinColumn(name = "afterSchool_id")
     private AfterSchool afterSchool;
 
-    private Long grade;
+    private Integer grade;
 
     public void mapping(AfterSchool afterSchool){
         this.afterSchool=afterSchool;
