+++ src/main/java/com/example/msgadminapi/dto/response/StatisticsResponseDto.java
@@ -9,7 +9,7 @@ import lombok.NoArgsConstructor;
 @AllArgsConstructor @NoArgsConstructor
 public class StatisticsResponseDto {
     @JsonProperty
-    private Long afterSchoolIdx;
+    private Integer afterSchoolIdx;
     @JsonProperty
     private String afterSchoolTitle;
     @JsonProperty
